+++ cmd/tko/main.go
@@ -8,8 +8,9 @@ import (
 	"strings"
 
 	"tko/internal/commands"
-	_ "tko/internal/commands/git" // register git handlers
-	_ "tko/internal/commands/ls"  // register ls handler
+	_ "tko/internal/commands/git"     // register git handlers
+	_ "tko/internal/commands/gobuild" // register go build handler
+	_ "tko/internal/commands/ls"      // register ls handler
 	"tko/internal/compress"
 	"tko/internal/hook"
 	"tko/internal/runner"